// src/proto-ts/packages/modules/common/proto/classpaths.ts
export enum Classpath {
  UNKNOWN = 0,
  BOOTCLASSPATH = 1,
  SYSTEMSERVERCLASSPATH = 2,
  DEX2OATBOOTCLASSPATH = 3,
  STANDALONE_SYSTEMSERVER_JARS = 4,
  UNRECOGNIZED = -1,
}
// ...
export interface Jar {
  /** Path on the filesystem for the jar. */
  path: string;
  /**
   * Environ classpath variable this jar belongs to.
   * Must be set to a known classpath.
   */
  classpath: Classpath;
  /**
   * Minimum API level required for the jar to be included on the classpath.
   * If the system's API level is lower than the value specified in this
   * attribute, the jar will not be included in the classpath.
   * Not setting this attribute implies the jar can be used on all API levels.
   */
  minSdkVersion: string;
  /**
   * Maximum API level that the jar file supports.
   * If the system's API level is higher than the value specified in this
   * attribute, the jar will not be included in the classpath.
   * Not setting this attribute implies unbound maximum.
   */
  maxSdkVersion: string;
}

/** Jars exported by a single partition. */
export interface ExportedClasspathsJars {
  /**
   * All jars that are exported as part of any classpath environ variable
   * (according to API level restrictions).
   * The relative order of the jars is preserved upon final merging.
   */
  jars: Jar[];
}

function createBaseJar(): Jar {
  return { path: "", classpath: 0, minSdkVersion: "", maxSdkVersion: "" };
}

export const Jar: MessageFns<Jar> = {
  encode(message: Jar, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.path !== "") {
      writer.uint32(10).string(message.path);
    }
    if (message.classpath !== 0) {
      writer.uint32(16).int32(message.classpath);
    }
    if (message.minSdkVersion !== "") {
      writer.uint32(26).string(message.minSdkVersion);
    }
    if (message.maxSdkVersion !== "") {
      writer.uint32(34).string(message.maxSdkVersion);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Jar {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseJar();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.path = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.classpath = reader.int32() as any;
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.minSdkVersion = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.maxSdkVersion = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Jar {
    return {
      path: isSet(object.path) ? globalThis.String(object.path) : "",
      classpath: isSet(object.classpath) ? classpathFromJSON(object.classpath) : 0,
      minSdkVersion: isSet(object.minSdkVersion)
        ? globalThis.String(object.minSdkVersion)
        : isSet(object.min_sdk_version)
        ? globalThis.String(object.min_sdk_version)
        : "",
      maxSdkVersion: isSet(object.maxSdkVersion)
        ? globalThis.String(object.maxSdkVersion)
        : isSet(object.max_sdk_version)
        ? globalThis.String(object.max_sdk_version)
        : "",
    };
  },

  toJSON(message: Jar): unknown {
    const obj: any = {};
    if (message.path !== "") {
      obj.path = message.path;
    }
    if (message.classpath !== 0) {
      obj.classpath = classpathToJSON(message.classpath);
    }
    if (message.minSdkVersion !== "") {
      obj.minSdkVersion = message.minSdkVersion;
    }
    if (message.maxSdkVersion !== "") {
      obj.maxSdkVersion = message.maxSdkVersion;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Jar>, I>>(base?: I): Jar {
    return Jar.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Jar>, I>>(object: I): Jar {
    const message = createBaseJar();
    message.path = object.path ?? "";
    message.classpath = object.classpath ?? 0;
    message.minSdkVersion = object.minSdkVersion ?? "";
    message.maxSdkVersion = object.maxSdkVersion ?? "";
    return message;
  },
};

function createBaseExportedClasspathsJars(): ExportedClasspathsJars {
  return { jars: [] };
}

export const ExportedClasspathsJars: MessageFns<ExportedClasspathsJars> = {
  encode(message: ExportedClasspathsJars, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.jars) {
      Jar.encode(v!, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ExportedClasspathsJars {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseExportedClasspathsJars();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.jars.push(Jar.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ExportedClasspathsJars {
    return { jars: globalThis.Array.isArray(object?.jars) ? object.jars.map((e: any) => Jar.fromJSON(e)) : [] };
  },

  toJSON(message: ExportedClasspathsJars): unknown {
    const obj: any = {};
    if (message.jars?.length) {
      obj.jars = message.jars.map((e) => Jar.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ExportedClasspathsJars>, I>>(base?: I): ExportedClasspathsJars {
    return ExportedClasspathsJars.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ExportedClasspathsJars>, I>>(object: I): ExportedClasspathsJars {
    const message = createBaseExportedClasspathsJars();
    message.jars = object.jars?.map((e) => Jar.fromPartial(e)) || [];
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}